src/oracle-node/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
web3==6.15.1
eth-account==0.10.0
pydantic==2.7.1
pyyaml==6.0.1
requests==2.32.3
